feat: moved app to single docker/script
