--- block-pvc-scanner/Dockerfile
@@ -1,11 +1,0 @@
-FROM python:3.8.7
-
-WORKDIR /usr/src/app
-
-COPY requirements.txt ./
-
-RUN pip install --no-cache-dir -r requirements.txt
-
-COPY scanner.py ./
-
-CMD [ "python", "./scanner.py" ]
