--- block-pvc-scanner/requirements.txt
@@ -1,2 +1,0 @@
-kubernetes==12.0.1
-prometheus_client==0.9.0
